refactor: extract parsing logic into own module
